More reasonable defaults
It's rare to use type info in a CSV. Usually it just makes the file unusable to other systems. Similarly, full DOM parsing of web responses adds overhead and sometimes even fails entirely (if IE defaults aren't set and the script is run non-interactively). These are lousy defaults, and should be opt-in.